Restructured code for readabiltiy
Wrapped pokemonlist inside an IIFE and assigend it to a new variable. added two function inside the IIFE to get the pokemonlist and to be able to add new pokemon to the pokemonlist.

The IIFE returns two methods

Lastly in order to retrieve and display the pokemonlist  the method getAll() was called